Conflicts:
	application/doctrine/data/fixtures/MigrationVersion.yml
	application/doctrine/data/fixtures/Role.yml
	application/modules/default/controllers/UserController.php
	application/modules/finding/controllers/DashboardController.php
	application/modules/finding/controllers/RemediationController.php
	library/Fisma/Cli/GenerateFindings.php
	public/javascripts/Fisma/User-min.js
	public/javascripts/Fisma/User.js
	public/javascripts/combined-min.js
	public/javascripts/combined.js

DashboardController.php: A query that was completely rewritten had previously had one of its table aliases
renamed. Now that it's been rewritten, however, that particular table alias isn't needed, so I took all of
the new stuff and threw away the old stuff.

GenerateFindings.php: Whitespace error involving a bug fix Josh B previously made. He added a blank line
at the same place that I added a new line of code. Resolved by replacing his blank line with my line of code.

Minified Stuff: Standard conflicts on minified JS resolved by re-minifying.

RemediationController.php: Conflict because Andrew moved init() to Finding_SummaryController on one branch
but added new print contexts on a different branch. Resolved by keeping the init() method in Remediation
Controller but getting rid of the lines that Andrew moved to the Finding_SummaryController.

Another conflict in RemediationController on the summary data stuff. Mark Ma modified the logic to use a
join table for "organization type" instead of hardcoded orgtype enum values, but this is the same code that
Andrew moved somewhere else.

Role.yml: New privileges added at the same place in the file. Resolved by keeping both sets of privileges.

User.js: Had a bug fix that overlapped my changes, but I had also fixed the same bug on my branch,
so I kept my side of the conflict and threw away the other side.

UserController.php: New code in the same place in the file. Resolved by keeping both additions of code.

Ran phpunit and php lint. Several tests fail but look like they were failing upstream, not failing due to
merge conflicts.

+++ application/modules/default/controllers/UserController.php
@@ -165,6 +165,19 @@ class UserController extends Fisma_Zend_Controller_Action_Object
 
             $this->_updatePocIndex($subject->id);
 
+            // Just send out email when create a new account or change password by admin user,
+            // and it does not sent out email when the root user changes his own password.
+            $actionName = strtolower($this->_request->getActionName());
+            if ('create' === $actionName) {
+                $mail = new Fisma_Zend_Mail();
+                $mail->sendAccountInfo($subject);
+            } else if ('edit' === $actionName
+                       && !empty($values['password'])
+                       && ('root' !== $subject->username || $this->_me->username !== 'root')) {
+                $mail = new Fisma_Zend_Mail();
+                $mail->sendPassword($subject);
+            }
+
             return $subject->id;
         } catch (Doctrine_Exception $e) {
             $conn->rollback();
@@ -321,28 +334,34 @@ class UserController extends Fisma_Zend_Controller_Action_Object
         $user = Doctrine::getTable('User')->find($this->_me->id);
 
         if ($this->_request->isPost()) {
-            //@todo check injection
-            $user->notifyFrequency = $this->_request->getParam('notify_frequency');
+            $notifyFrequency = $this->_request->getParam('notify_frequency');
 
             $postEvents = $this->_request->getPost('existEvents');
-            try {
-                Doctrine_Manager::connection()->beginTransaction();
-                $modified = $user->getModified();
+            if (Inspekt::isInt($notifyFrequency)) {
+                try {
+                    Doctrine_Manager::connection()->beginTransaction();
+                    $modified = $user->getModified();
 
-                $user->unlink('Events');
+                    $user->unlink('Events');
 
-                if (!empty($postEvents)) {
-                    $user->link('Events', $postEvents);
+                    if (!empty($postEvents)) {
+                        $user->link('Events', $postEvents);
+                    }
+
+                    $user->notifyFrequency = $notifyFrequency;
+                    $user->save();
+                    Doctrine_Manager::connection()->commit();
+
+                    $message = "Notification events modified successfully";
+                    $model   = 'notice';
+                } catch (Doctrine_Exception $e) {
+                    Doctrine_Manager::connection()->rollback();
+                    $message = $e->getMessage();
+                    $model   = 'warning';
                 }
-
-                $user->save();
-                Doctrine_Manager::connection()->commit();
-
-                $message = "Notification events modified successfully";
-                $model   = 'notice';
-            } catch (Doctrine_Exception $e) {
-                Doctrine_Manager::connection()->rollback();
-                $message = $e->getMessage();
+            } else {
+                /** @todo English */
+                $message = "Notify Frequency: '$notifyFrequency' is not a valid value";
                 $model   = 'warning';
             }
             $this->view->priorityMessenger($message, $model);
